[improvement] Swapper Implementation improvements #44
Closes #44
