+++ .travis.yml
@@ -1,5 +1,4 @@
 language: python
-sudo: false
 dist: bionic
 
 cache: pip
@@ -26,7 +25,8 @@ install:
 script:
   - ./run-qa-checks
   - coverage run --source=openwisp_ipam runtests.py
-  - SAMPLE_APP=1 ./runtests.py
+  # SAMPLE tests do not influence coverage, so we can speed up tests with --parallel
+  - SAMPLE_APP=1 ./runtests.py --parallel --keepdb
 
 after_success:
   - coveralls
