+++ openwisp_ipam/tests/test_users_integration.py
@@ -1,0 +1,8 @@
+from openwisp_users.tests.test_admin import TestUsersAdmin
+
+
+class TestUsersIntegration(TestUsersAdmin):
+    pass
+
+
+del TestUsersAdmin
