+++ tests/openwisp2/sample_ipam/admin.py
@@ -1,16 +1,5 @@
-import swapper
-from django.contrib import admin
-from openwisp_ipam.base.admin import AbstractIpAddressAdmin, AbstractSubnetAdmin
+from openwisp_ipam.admin import IpAddressAdmin, SubnetAdmin  # noqa
 
-IpAddress = swapper.load_model("sample_ipam", "IpAddress")
-Subnet = swapper.load_model("sample_ipam", "Subnet")
-
-
-@admin.register(IpAddress)
-class IPAddressAdmin(AbstractIpAddressAdmin):
-    pass
-
-
-@admin.register(Subnet)
-class SubnetAdmin(AbstractSubnetAdmin):
-    app_name = "sample_ipam"
+# Monkey Patching Subnet Admin to change the name of the
+# app_label used. Change this value to your <app_label>.
+SubnetAdmin.app_label = 'sample_ipam'
