+++ tests/openwisp2/sample_ipam/management/commands/export_subnet.py
@@ -1,6 +1,5 @@
 import swapper
-
-from . import BaseExportSubnetCommand
+from openwisp_ipam.management.commands import BaseExportSubnetCommand
 
 
 class Command(BaseExportSubnetCommand):
