+++ tests/openwisp2/sample_ipam/models.py
@@ -3,6 +3,12 @@ from openwisp_ipam.base.models import AbstractIpAddress, AbstractSubnet
 
 
 class DetailsModel(models.Model):
+    """
+    This field serves no purpose, it only serves as an example
+    for extending models and used for testing purposes.
+    It will be inherited by all the models.
+    """
+
     details = models.CharField(max_length=64, blank=True, null=True)
 
     class Meta:
